Move DPM support protocols to DPM module
