+++ src/app/services/dpm_proof_pack_service.py
@@ -11,11 +11,12 @@ from app.contracts.dpm_proof_packs import (
     DpmProofPackMemoRequest,
     DpmProofPackSupportability,
 )
+from app.services.dpm_client_protocols import DpmProofPackClient
 from app.services.lotus_ai_workflow import (
     build_workflow_pack_task_request,
     require_lotus_ai_client,
 )
-from app.services.upstream_client_protocols import DpmProofPackClient, LotusAiWorkflowClient
+from app.services.upstream_client_protocols import LotusAiWorkflowClient
 from app.services.upstream_envelope import (
     build_upstream_status_gateway_envelope,
     raise_product_safe_service_error,
