+++ src/app/services/upstream_client_protocols.py
@@ -1,69 +1,6 @@
 from typing import Any, Protocol
 
 
-class DpmConstructionClient(Protocol):
-    async def generate_construction_alternative_set(
-        self,
-        *,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_construction_alternative_set(
-        self,
-        *,
-        alternative_set_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def select_construction_alternative(
-        self,
-        *,
-        alternative_set_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
-class DpmProofPackClient(Protocol):
-    async def generate_proof_pack(
-        self,
-        *,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proof_pack(
-        self,
-        *,
-        proof_pack_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proof_pack_markdown(
-        self,
-        *,
-        proof_pack_id: str,
-        correlation_id: str,
-    ) -> tuple[int, str, dict[str, Any]]: ...
-
-    async def get_proof_pack_report_input(
-        self,
-        *,
-        proof_pack_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proof_pack_ai_evidence_input(
-        self,
-        *,
-        proof_pack_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
 class LotusAiWorkflowClient(Protocol):
     async def execute_workflow_pack(
         self,
